enable install through go install
