+++ Makefile
@@ -32,7 +32,7 @@ deps:
 .PHONY: build
 build: 
 	rm -f ${GOPATH}/bin/applier
-	go install ./cmd/applier.go
+	go install ./applier/applier.go
 
 .PHONY: build-bin
 build-bin: doc-help
@@ -40,13 +40,13 @@ build-bin: doc-help
 	zip -q docs/help.zip -j docs/help/*
 	@rm -rf bin
 	@mkdir -p bin
-	GOOS=darwin GOARCH=amd64 go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./cmd/applier.go && tar -czf bin/applier_darwin_amd64.tar.gz LICENSE -C bin/ applier
-	GOOS=darwin GOARCH=arm64 go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./cmd/applier.go && tar -czf bin/applier_darwin_arm64.tar.gz LICENSE -C bin/ applier
-	GOOS=linux GOARCH=amd64 go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./cmd/applier.go && tar -czf bin/applier_linux_amd64.tar.gz LICENSE -C bin/ applier 
-	GOOS=linux GOARCH=arm64 go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./cmd/applier.go && tar -czf bin/applier_linux_arm64.tar.gz LICENSE -C bin/ applier 
-	GOOS=linux GOARCH=ppc64le go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./cmd/applier.go && tar -czf bin/applier_linux_ppc64le.tar.gz LICENSE -C bin/ applier 
-	GOOS=linux GOARCH=s390x go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./cmd/applier.go && tar -czf bin/applier_linux_s390x.tar.gz LICENSE -C bin/ applier
-	GOOS=windows GOARCH=amd64 go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier.exe ./cmd/applier.go && zip -q bin/applier_windows_amd64.zip LICENSE -j bin/applier.exe
+	GOOS=darwin GOARCH=amd64 go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./applier/applier.go && tar -czf bin/applier_darwin_amd64.tar.gz LICENSE -C bin/ applier
+	GOOS=darwin GOARCH=arm64 go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./applier/applier.go && tar -czf bin/applier_darwin_arm64.tar.gz LICENSE -C bin/ applier
+	GOOS=linux GOARCH=amd64 go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./applier/applier.go && tar -czf bin/applier_linux_amd64.tar.gz LICENSE -C bin/ applier 
+	GOOS=linux GOARCH=arm64 go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./applier/applier.go && tar -czf bin/applier_linux_arm64.tar.gz LICENSE -C bin/ applier 
+	GOOS=linux GOARCH=ppc64le go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./applier/applier.go && tar -czf bin/applier_linux_ppc64le.tar.gz LICENSE -C bin/ applier 
+	GOOS=linux GOARCH=s390x go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier ./applier/applier.go && tar -czf bin/applier_linux_s390x.tar.gz LICENSE -C bin/ applier
+	GOOS=windows GOARCH=amd64 go build -ldflags="-s -w" -gcflags=-trimpath=x/y  -o bin/applier.exe ./applier/applier.go && zip -q bin/applier_windows_amd64.zip LICENSE -j bin/applier.exe
 
 .PHONY: release
 release:
