+++ pkg/asset/embedFSReader.go
@@ -12,15 +12,6 @@ import (
 	"github.com/ghodss/yaml"
 )
 
-type ScenarioReader interface {
-	//Retrieve an asset from the data source
-	Asset(templatePath string) ([]byte, error)
-	//List all available assets in the data source
-	AssetNames(excluded []string) ([]string, error)
-	ExtractAssets(prefix, dir string, excluded []string) error
-	ToJSON(b []byte) ([]byte, error)
-}
-
 type ScenarioResourcesReader struct {
 	files *embed.FS
 }
